master/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
PyJWT[crypto]==2.9.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<5.0
cryptography==42.0.0
httpx==0.27.0
websockets==12.0
pyyaml==6.0.1
python-multipart==0.0.9
pytest==8.2.0
pytest-asyncio==0.23.6
